Создана фабрика вопросов. Модели вопросов вынесены в отдельные файлы.
