+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -5,4 +5,10 @@
 //  Created by Дмитрий Ивашинин on 21.11.2023.
 //
 
-import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+  let image: UIImage
+  let question: String
+  let questionNumber: String
+}
